requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
openai==1.43.0
pydantic-settings==2.4.0
pytest
pytest-asyncio
pymupdf==1.24.9
langchain==0.2.15
python-multipart==0.0.9
langchain-community==0.2.15
langchain-openai==0.1.23
python-dotenv==1.0.1
langchain-chroma==0.1.3
chromadb==0.5.3